vllm_logger/pyproject.toml
[project]
name = "vllm-logger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nvidia-ml-py>=13.590.44",
    "httpx>=0.27.0",
    "tqdm>=4.67.1",
    "requests>=2.32.5",
    "python-dotenv>=1.2.1",
]
